update ingest_vring
for some vendors (e.g., vhost), the device may read part of an
descriptor instead full of them, which causes re-entrance of the
generation of the same desc, resulting errors. to fix this, we need to
identify whether a desc has been generated yet, by recording all
generated desc indexes. if generated, we skip DMA hook and return.