update switch value generation machanism
In the old version, switch values are generated according to a biased
distribution only when we need to introduce new bytes into inputs. If
bytes of the switch value are already in the input, we would not
generate it.

Now, we will re-generate switch value if it is larger than 0xffff