the memory range of gcov_info should be persistent. Including gcov_info->functions, *gcov_info->functions[i], and gcov_ctr_info->values.
Hyperpill assumes each page has at most one range to be persistent. This is not true for
gcov_info, which has multiple ranges (one for gcov_info itself, one for functions, and ones for values).
To fix this, we corresponds each page with a vector of ranges to be persistent. All ranges should keep in order.